+++ spec/www/spec/sqlite-version-test.js
@@ -102,7 +102,7 @@ var mytests = function() {
               expect(rs).toBeDefined();
               expect(rs.rows).toBeDefined();
               expect(rs.rows.length).toBe(1);
-              expect(rs.rows.item(0).myResult).toBe('3.32.3');
+              expect(rs.rows.item(0).myResult).toBe('3.40.0');
 
               // Close (plugin only) & finish:
               (isWebSql) ? done() : db.close(done, done);